% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop.R
\name{loop}
\alias{loop}
\title{Loop for computate all entropy metrics outputs
Contingence table of normalized mutual information theory coefficients}
\usage{
loop(x)
}
\arguments{
\item{x}{is the working directory}
}
\description{
Loop for computate all entropy metrics outputs
Contingence table of normalized mutual information theory coefficients
}
