% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearsontable.R
\name{pearsontable}
\alias{pearsontable}
\title{Loop for computate all Pearson correlation outputs
Contingence table of Pearson coefficients}
\usage{
pearsontable(x)
}
\arguments{
\item{x}{is the source data csv file}
}
\description{
Loop for computate all Pearson correlation outputs
Contingence table of Pearson coefficients
}
