% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap2.R
\name{heatmap2}
\alias{heatmap2}
\title{Shannon entropy heatmap}
\usage{
heatmap2(z)
}
\arguments{
\item{z}{is the entropy metrics outputs matrix/data frame (computated previously with the loop function) file}
}
\description{
Contingence table heatmap of normalized mutual information theory coefficients with the Chi2 p-value. Studied variables can be binary or quantitative.
}
\examples{
## Entropy outputs coumputation and visualization
### Computation
loop(docs_phenotype_file_1,1,8)
### Visualization on a heatmap
entropy_outputs <- readr::read_csv('entropy_outputs.csv')
heatmap2(entropy_outputs)
}
\author{
Edouard Lansiaux, Philippe Pierre Pébaÿ
}
