% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap2.R
\name{heatmap2}
\alias{heatmap2}
\title{Contingence table heatmap
Contingence table of normalized mutual information theory coefficients}
\usage{
heatmap2(z)
}
\arguments{
\item{z}{is the entropy metrics outputs csv file}
}
\description{
Contingence table heatmap
Contingence table of normalized mutual information theory coefficients
}
