% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop.R
\name{loop}
\alias{loop}
\title{Loop for computate all entropy metrics outputs
Contingence table of normalized mutual information theory coefficients}
\usage{
loop(x, fn, m, n)
}
\arguments{
\item{x}{is the working directory}

\item{fn}{is the path to source data csv file}

\item{m}{is the first studied X variable column number}

\item{n}{is the studied variables number}
}
\description{
Loop for computate all entropy metrics outputs
Contingence table of normalized mutual information theory coefficients
}
