% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearsontable.R
\name{pearsontable}
\alias{pearsontable}
\title{Pearson correlation heatmap}
\usage{
pearsontable(x)
}
\arguments{
\item{x}{is the source data matrix/data frame}
}
\description{
Contingence table heatmap of Pearson coefficients and p-values. Studied variables can be binary or quantitative.Concerning Pearson correlation, it is a commonly formulated criticism that one may not establish a linear correlation between a series of quantitative variables and another one of qualitative variables . However, this concern is misguided in the case of dichotomous variables (i.e., taking binary values), for this correlation can be legitimately established using the point biserial correlation coefficient.
}
\examples{
## Pearson correlation computation and visualization
pearsontable(docs_phenotype_file_1)
}
\references{
1/ Lev, J. The point biserial coefficient of correlation. Ann. Math. Stat. 20, DOI: 10.1214/aoms/1177730103 (1949). 2/ Tate, R. Correlation between a discrete and a continuous variable. point-biserial correlation. The Annals Math. Stat. 25,DOI: 10.1214/aoms/1177728730 (1954) 3/ Kornbrot, D. Point biserial correlation. In Encyclopedia of Statistics in Behavioral Science, DOI: 10.1002/0470013192.bsa485 (American Cancer Society, 2005)
}
\author{
Joachim Forget, Jean-Luc Caut
}
